Rework metric provider processing mechanism and removed storing to DB / run_id
run_id is now added in segmented helper metric_importer. Run ID is a concept the metric providers should not be aware of and
is thus handled now externally.
Furthermore the processing mechanism in the providers does now not allow overloading of externally callable functions
in read_metrics(). New private sub functions are created that go through different steps of reading, parsing, validation and adding data.